XAP-12508,XAP-13033
changed ReplicationPacketData weight to type int,
added predictedWeight for monitoring and unit
test

+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/cluster/node/impl/backlog/globalorder/GlobalOrderOperationPacket.java
@@ -87,7 +87,7 @@ public class GlobalOrderOperationPacket
         }
     }
 
-    public long getWeight() {
+    public int getWeight() {
         return _data.getWeight();
     }
