+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/cluster/node/impl/packets/data/IReplicationPacketData.java
@@ -66,8 +66,8 @@ public interface IReplicationPacketData<T extends IReplicationPacketEntryData>
      */
     boolean isFromGateway();
 
-    void setWeight(long weight);
+    void setWeight(int weight);
 
-    long getWeight();
+    int getWeight();
 
 }
