+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/cluster/node/impl/packets/data/operations/AbstractReplicationPacketSingleEntryData.java
@@ -60,7 +60,7 @@ public abstract class AbstractReplicationPacketSingleEntryData
     private static final int FLAGS_GATEWAY = 1 << 0;
     private static final int FLAGS_BLOBSTORE_BULK = 1 << 1;
 
-    private transient long _weight;
+    private transient int _weight;
     //Externalizable
     public AbstractReplicationPacketSingleEntryData() {
     }
@@ -367,12 +367,12 @@ public abstract class AbstractReplicationPacketSingleEntryData
     }
 
     @Override
-    public void setWeight(long weight) {
+    public void setWeight(int weight) {
         _weight = weight;
     }
 
     @Override
-    public long getWeight() {
+    public int getWeight() {
         return _weight;
     }
 }
